+++ R/R/plot.hBayesDM.R
@@ -5,8 +5,8 @@
 #' @param x Model output of class hBayesDM
 #' @param type Character value that specifies the plot type. Options are: "dist", "trace", or "simple". Defaults to "dist".
 #' @param ncols Integer value specifying how many plots there should be per row. Defaults to the number of parameters.
-#' @param fontSize Integer value specifying the size of the font used for plotting. Defaults to 10.
-#' @param binSize Integer value specifying how wide the bars on the histogram should be. Defaults to 30.
+#' @param font_size Integer value specifying the size of the font used for plotting. Defaults to 10.
+#' @param bin_size Integer value specifying how wide the bars on the histogram should be. Defaults to 30.
 #' @param ... Additional arguments to be passed on
 #'
 #' @importFrom bayesplot mcmc_trace mcmc_intervals
@@ -17,8 +17,8 @@
 plot.hBayesDM <- function(x        = NULL,
                           type     = "dist",
                           ncols    = NULL,
-                          fontSize = NULL,
-                          binSize  = NULL,
+                          font_size = NULL,
+                          bin_size  = NULL,
                           ...) {
 
   # cmdstanr stores VB results in CmdStanVB; MCMC in CmdStanMCMC.
@@ -33,18 +33,18 @@ plot.hBayesDM <- function(x        = NULL,
   if (grepl(pattern = "lba", x = x$model)) {
     numPars <- 4
   } else {
-    numPars <- dim(x$allIndPars)[2] - 1
+    numPars <- dim(x$all_ind_pars)[2] - 1
   }
 
-  parNames <- names(x$parVals)[1:numPars]
+  parNames <- names(x$par_vals)[1:numPars]
 
   if (type == "dist") {
     source(file = system.file("plotting", "plot_functions.R", package = "hBayesDM"),
            local = TRUE)
     eval(parse(text = paste0("plot_", x$model, "(obj = x",
-                             ", fontSize = ", fontSize,
+                             ", font_size = ", font_size,
                              ", ncols = ", ncols,
-                             ", binSize = ", binSize, ")")))
+                             ", bin_size = ", bin_size, ")")))
     invisible()
   } else if (type == "trace") {
     bayesplot::mcmc_trace(x$fit$draws(parNames),